组织规则引擎相关的结果识别代码，复用相同代码
